+++ src/index.ts
@@ -6,38 +6,41 @@
  */
 
 import { IconBack } from './icons/IconBack';
-import { IconQiYeBan } from './icons/IconQiYeBan';
-import { IconSheZhi } from './icons/IconSheZhi';
-import { IconTianJia } from './icons/IconTianJia';
-import { IconWangLuoJianSuo } from './icons/IconWangLuoJianSuo';
-import { IconXiaZaiZhuShou } from './icons/IconXiaZaiZhuShou';
-import { IconXiTongPeiZhi } from './icons/IconXiTongPeiZhi';
-import { IconXiTongPeiZhi1 } from './icons/IconXiTongPeiZhi1';
-import { IconYunYingGuanLi } from './icons/IconYunYingGuanLi';
+import { IconCheck } from './icons/IconCheck';
+import { IconGuanBi } from './icons/IconGuanBi';
+import { IconGuanBiXiao } from './icons/IconGuanBiXiao';
+import { IconLianJie } from './icons/IconLianJie';
+import { IconSearch } from './icons/IconSearch';
+import { IconTongZhi } from './icons/IconTongZhi';
+import { IconWenDang } from './icons/IconWenDang';
+import { IconZhanKai } from './icons/IconZhanKai';
+import { IconZhiShiKu } from './icons/IconZhiShiKu';
 
 export { IconBack, type IconBackProps } from './icons/IconBack';
-export { IconQiYeBan, type IconQiYeBanProps } from './icons/IconQiYeBan';
-export { IconSheZhi, type IconSheZhiProps } from './icons/IconSheZhi';
-export { IconTianJia, type IconTianJiaProps } from './icons/IconTianJia';
-export { IconWangLuoJianSuo, type IconWangLuoJianSuoProps } from './icons/IconWangLuoJianSuo';
-export { IconXiaZaiZhuShou, type IconXiaZaiZhuShouProps } from './icons/IconXiaZaiZhuShou';
-export { IconXiTongPeiZhi, type IconXiTongPeiZhiProps } from './icons/IconXiTongPeiZhi';
-export { IconXiTongPeiZhi1, type IconXiTongPeiZhi1Props } from './icons/IconXiTongPeiZhi1';
-export { IconYunYingGuanLi, type IconYunYingGuanLiProps } from './icons/IconYunYingGuanLi';
+export { IconCheck, type IconCheckProps } from './icons/IconCheck';
+export { IconGuanBi, type IconGuanBiProps } from './icons/IconGuanBi';
+export { IconGuanBiXiao, type IconGuanBiXiaoProps } from './icons/IconGuanBiXiao';
+export { IconLianJie, type IconLianJieProps } from './icons/IconLianJie';
+export { IconSearch, type IconSearchProps } from './icons/IconSearch';
+export { IconTongZhi, type IconTongZhiProps } from './icons/IconTongZhi';
+export { IconWenDang, type IconWenDangProps } from './icons/IconWenDang';
+export { IconZhanKai, type IconZhanKaiProps } from './icons/IconZhanKai';
+export { IconZhiShiKu, type IconZhiShiKuProps } from './icons/IconZhiShiKu';
 
 /**
  * Object containing all icon components
  */
 export const allIcons = {
   IconBack,
-  IconQiYeBan,
-  IconSheZhi,
-  IconTianJia,
-  IconWangLuoJianSuo,
-  IconXiaZaiZhuShou,
-  IconXiTongPeiZhi,
-  IconXiTongPeiZhi1,
-  IconYunYingGuanLi
+  IconCheck,
+  IconGuanBi,
+  IconGuanBiXiao,
+  IconLianJie,
+  IconSearch,
+  IconTongZhi,
+  IconWenDang,
+  IconZhanKai,
+  IconZhiShiKu
 } as const;
 
 /**