+++ CHANGELOG.md
@@ -6,6 +6,78 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.17] - 2026-01-30
+
+更新/新增 10 个图标：back, check, search, 关闭, 关闭-小
+
+### Added
+- IconZhiShiKu
+- IconTongZhi
+- IconLianJie
+- IconGuanBi
+- IconGuanBiXiao
+- IconZhanKai
+- IconWenDang
+- IconCheck
+- IconSearch
+- IconBack
+
+### Modified
+- None
+
+### Removed
+- None
+## [0.0.16] - 2026-01-30
+
+更新/新增 10 个图标：back, check, search, 关闭, 关闭-小
+
+### Added
+- IconTongZhi
+- IconLianJie
+- IconZhanKai
+- IconWenDang
+- IconZhiShiKu
+- IconGuanBi
+- IconGuanBiXiao
+- IconSearch
+- IconBack
+- IconCheck
+
+### Modified
+- None
+
+### Removed
+- None
+## [0.0.15] - 2026-01-30
+
+同步 10 个图标：back, check, search, 关闭, 关闭-小
+
+### Added
+- IconTongZhi
+- IconLianJie
+- IconWenDang
+- IconZhiShiKu
+- IconGuanBi
+- IconGuanBiXiao
+- IconZhanKai
+- IconSearch
+- IconBack1
+- IconCheck
+- IconYunYingGuanLi
+- IconXiTongPeiZhi
+- IconXiTongPeiZhi1
+- IconXiaZaiZhuShou
+- IconQiYeBan
+- IconSheZhi
+- IconTianJia
+- IconWangLuoJianSuo
+- IconBack
+
+### Modified
+- None
+
+### Removed
+- None
 ## [0.0.13] - 2026-01-30
 
 同步 5 个图标：下载助手, 企业版, 系统配置, 系统配置, 运营管理 图标